Refactor FFT code to have switchable backends
This allows us to add a PipeWire reader in later, which is necessary due
to Flatpak preventing us from reading namedpipes.
