+++ Cargo.toml
@@ -43,6 +43,7 @@ libc = "0.2.169"
 spectrum-analyzer = "=0.5.2"
 ashpd = "0.10.2"
 tokio = { version = "1", features = ["rt-multi-thread"] }
+pipewire = "0.8.0"
 
 [dependencies.gtk]
 package = "gtk4"
