+++ src/player/controller.rs
@@ -1,10 +1,6 @@
 extern crate mpd;
 use crate::{
-    application::EuphonicaApplication,
-    cache::{Cache, CacheState},
-    client::{ClientState, ConnectionState, MpdWrapper},
-    common::{AlbumInfo, QualityGrade, Song},
-    utils::{prettify_audio_format, settings_manager},
+    application::EuphonicaApplication, cache::{Cache, CacheState}, client::{ClientState, ConnectionState, MpdWrapper}, common::{AlbumInfo, QualityGrade, Song}, player::fft_backends::fifo::FifoFftBackend, utils::{prettify_audio_format, settings_manager}
 };
 use async_lock::OnceCell as AsyncOnceCell;
 use mpris_server::{
@@ -24,19 +20,12 @@ use mpd::{
     ReplayGain, SaveMode, Subsystem,
 };
 use std::{
-    cell::{Cell, OnceCell, RefCell},
-    ops::Deref,
-    path::PathBuf,
-    rc::Rc,
-    str::FromStr,
-    sync::{
-        atomic::{AtomicBool, Ordering},
+    cell::{Cell, OnceCell, RefCell}, ops::Deref, path::PathBuf, rc::Rc, sync::{
         Arc, Mutex, OnceLock,
-    },
-    thread,
-    time::Duration,
-    vec::Vec,
+    }, vec::Vec
 };
+
+use super::fft_backends::backend::{FftStatus, FftBackend};
 
 #[derive(Clone, Copy, Debug, glib::Enum, PartialEq, Default)]
 #[enum_type(name = "EuphonicaPlaybackState")]
@@ -65,15 +54,6 @@ pub enum PlaybackFlow {
     Repeat,       // Loops through the queue
     Single,       // Plays one song then stops & waits for click to play next song
     RepeatSingle, // Loops current song
-}
-
-#[derive(Clone, Copy, Debug, glib::Enum, PartialEq, Default)]
-#[enum_type(name = "EuphonicaFftStatus")]
-pub enum FftStatus {
-    #[default]
-    Invalid,
-    ValidNotReading, // due to visualiser not being run
-    Reading,
 }
 
 impl FftStatus {
@@ -218,9 +198,8 @@ mod imp {
         pub app: OnceCell<EuphonicaApplication>,
         pub supports_playlists: Cell<bool>,
         // For receiving frequency levels from FFT thread
-        pub fft_stop: Arc<AtomicBool>,
+        pub fft_backend: RefCell<Rc<dyn FftBackend>>,
         pub fft_data: Arc<Mutex<(Vec<f32>, Vec<f32>)>>, // Binned magnitudes, in stereo
-        pub fft_handle: RefCell<Option<gio::JoinHandle<()>>>,
         pub use_visualizer: Cell<bool>,
     }
 
@@ -230,6 +209,7 @@ mod imp {
         type Type = super::Player;
 
         fn new() -> Self {
+
             Self {
                 state: Cell::new(PlaybackState::Stopped),
                 position: Cell::new(0.0),
@@ -252,7 +232,7 @@ mod imp {
                 mpris_server: AsyncOnceCell::new(),
                 mpris_enabled: Cell::new(false),
                 app: OnceCell::new(),
-                fft_stop: Arc::new(AtomicBool::new(false)),
+                fft_backend: RefCell::new(Rc::new(FifoFftBackend::default())),
                 fft_data: Arc::new(Mutex::new((
                     vec![
                         0.0;
@@ -267,11 +247,16 @@ mod imp {
                             .uint("visualizer-spectrum-bins") as usize
                     ],
                 ))),
-                fft_handle: RefCell::new(None),
                 use_visualizer: Cell::new(false),
             }
         }
     }
+
+impl Default for Player {
+    fn default() -> Self {
+        Self::new()
+    }
+}
 
     impl ObjectImpl for Player {
         fn constructed(&self) {
@@ -289,13 +274,7 @@ mod imp {
         }
 
         fn dispose(&self) {
-            glib::MainContext::default().spawn_local(clone!(
-                #[weak(rename_to = this)]
-                self,
-                async move {
-                    this.obj().maybe_stop_fft_thread().await;
-                }
-            ));
+            self.obj().maybe_stop_fft_thread();
         }
 
         fn properties() -> &'static [ParamSpec] {
@@ -418,13 +397,7 @@ mod imp {
                         } else {
                             // Visualiser turned off. FFT thread should
                             // have stopped by itself. Join & yeet handle.
-                            glib::MainContext::default().spawn_local(clone!(
-                                #[weak(rename_to = this)]
-                                self,
-                                async move {
-                                    this.obj().maybe_stop_fft_thread().await;
-                                }
-                            ));
+                            self.obj().maybe_stop_fft_thread();
                         }
                     }
                 }
@@ -485,143 +458,19 @@ impl Player {
     // 2. Perform FFT & extrapolate to the marker frequencies.
     // 3. Send results back to main thread via the async channel.
     fn maybe_start_fft_thread(&self) {
-        let stop_flag = self.imp().fft_stop.clone();
-        stop_flag.store(false, Ordering::Relaxed);
         let output = self.imp().fft_data.clone();
-        let should_start: bool;
-        {
-            should_start = self.imp().fft_handle.borrow().as_ref().is_none();
-        }
-        if should_start {
-            let fft_handle = gio::spawn_blocking(move || {
-                let settings = settings_manager();
-                let player_settings = settings.child("player");
-                // Will require starting a new thread to account for path and format changes
-                if let Ok(format) = AudioFormat::from_str(
-                    settings.child("client").string("mpd-fifo-format").as_str(),
-                ) {
-                    // These settings require a restart
-                    let n_samples = player_settings.uint("visualizer-fft-samples") as usize;
-                    let n_bins = player_settings.uint("visualizer-spectrum-bins") as usize;
-                    if let Ok(mut reader) = super::fft::try_open_pipe(
-                        settings.child("client").string("mpd-fifo-path").as_str(),
-                        &format,
-                        n_samples,
-                    ) {
-                        // Allocate the following once only
-                        let mut fft_buf_left: Vec<f32> = vec![0.0; n_samples];
-                        let mut fft_buf_right: Vec<f32> = vec![0.0; n_samples];
-                        let mut curr_step_left: Vec<f32> = vec![0.0; n_bins];
-                        let mut curr_step_right: Vec<f32> = vec![0.0; n_bins];
-                        'outer: loop {
-                            if stop_flag.load(Ordering::Relaxed)
-                                || !settings.child("ui").boolean("use-visualizer")
-                            {
-                                break 'outer;
-                            }
-                            // These should be applied on-the-fly
-                            let bin_mode =
-                                if player_settings.boolean("visualizer-spectrum-use-log-bins") {
-                                    super::fft::BinMode::Logarithmic
-                                } else {
-                                    super::fft::BinMode::Linear
-                                };
-                            let fps = player_settings.uint("visualizer-fps") as f32;
-                            let min_freq =
-                                player_settings.uint("visualizer-spectrum-min-hz") as f32;
-                            let max_freq =
-                                player_settings.uint("visualizer-spectrum-max-hz") as f32;
-                            let curr_step_weight = player_settings
-                                .double("visualizer-spectrum-curr-step-weight")
-                                as f32;
-                            match super::fft::get_stereo_pcm(
-                                &mut fft_buf_left,
-                                &mut fft_buf_right,
-                                &mut reader,
-                                &format,
-                                fps,
-                                true,
-                            ) {
-                                Ok(()) => {
-                                    // Compute outside of mutex lock please
-                                    super::fft::get_magnitudes(
-                                        &format,
-                                        &mut fft_buf_left,
-                                        &mut curr_step_left,
-                                        n_bins as u32,
-                                        bin_mode,
-                                        min_freq,
-                                        max_freq,
-                                    );
-                                    super::fft::get_magnitudes(
-                                        &format,
-                                        &mut fft_buf_right,
-                                        &mut curr_step_right,
-                                        n_bins as u32,
-                                        bin_mode,
-                                        min_freq,
-                                        max_freq,
-                                    );
-                                    // Replace last frame
-                                    if let Ok(mut output_lock) = output.lock() {
-                                        if output_lock.0.len() != n_bins
-                                            || output_lock.1.len() != n_bins
-                                        {
-                                            output_lock.0.clear();
-                                            output_lock.1.clear();
-                                            for _ in 0..n_bins {
-                                                output_lock.0.push(0.0);
-                                                output_lock.1.push(0.0);
-                                            }
-                                        }
-                                        for i in 0..n_bins {
-                                            output_lock.0[i] = curr_step_left[i] * curr_step_weight
-                                                + output_lock.0[i] * (1.0 - curr_step_weight);
-                                            output_lock.1[i] = curr_step_right[i]
-                                                * curr_step_weight
-                                                + output_lock.1[i] * (1.0 - curr_step_weight);
-                                        }
-                                        // println!("FFT L: {:?}\tR: {:?}", &output_lock.0, &output_lock.1);
-                                    } else {
-                                        panic!("Unable to lock FFT data mutex");
-                                    }
-                                }
-                                Err(e) => match e.kind() {
-                                    std::io::ErrorKind::UnexpectedEof
-                                    | std::io::ErrorKind::WouldBlock => {}
-                                    _ => {
-                                        println!("FFT ERR: {:?}", &e);
-                                        break 'outer;
-                                    }
-                                },
-                            }
-                            thread::sleep(Duration::from_millis((1000.0 / fps).floor() as u64));
-                        }
-                    }
-                }
-            });
-            self.imp().fft_handle.replace(Some(fft_handle));
+        if let Ok(()) = self.imp().fft_backend.borrow().start(output) {
             self.notify("fft-status");
         }
     }
 
-    async fn maybe_stop_fft_thread(&self) {
-        if let Some(handle) = self.imp().fft_handle.take() {
-            self.imp().fft_stop.store(true, Ordering::Relaxed);
-            let _ = handle.await;
-            self.notify("fft-status");
-        }
+    fn maybe_stop_fft_thread(&self) {
+        self.imp().fft_backend.borrow().stop();
     }
 
-    pub fn reconnect_fifo(&self) {
-        glib::MainContext::default().spawn_local(clone!(
-            #[weak(rename_to = this)]
-            self,
-            async move {
-                this.maybe_stop_fft_thread().await;
-                this.maybe_start_fft_thread();
-            }
-        ));
+    pub fn restart_fft_thread(&self) {
+        let output = self.imp().fft_data.clone();
+        self.imp().fft_backend.borrow().clone().restart(output);
     }
 
     pub fn fft_data(&self) -> Arc<Mutex<(Vec<f32>, Vec<f32>)>> {
@@ -1194,20 +1043,7 @@ impl Player {
     }
 
     pub fn fft_status(&self) -> FftStatus {
-        if self.imp().fft_handle.borrow().as_ref().is_some() {
-            return FftStatus::Reading;
-        } else {
-            // Try to open a new reader
-            if let Ok(_) = super::fft::open_named_pipe_readonly(
-                settings_manager()
-                    .child("client")
-                    .string("mpd-fifo-path")
-                    .as_str(),
-            ) {
-                return FftStatus::ValidNotReading;
-            }
-            return FftStatus::Invalid;
-        }
+        self.imp().fft_backend.borrow().status()
     }
 
     pub fn format_desc(&self) -> Option<String> {
